[Architecture] Re-designing website architecture.
