+++ app.ini
@@ -1,0 +1,17 @@
+[uwsgi]
+; Load the callable instance 'app' from module 'app.py'.
+module = app
+callable = app
+
+; There's no magic number to set the number of processes, see:
+; http://uwsgi-docs.readthedocs.io/en/latest/ThingsToKnow.html
+master = true
+processes = 4
+
+; Run server on port 5000 with socket communications.
+socket = :5000
+
+; Where the uswgi related log goes to and change permission so that
+; we can read the file outside container with user permission.
+logto = /home/richo.tw/log/uwsgi.log
+logfile-chmod = 664
