+++ conf.d/richo.tw.conf
@@ -1,0 +1,10 @@
+server {
+    listen 80;
+
+    error_log /var/log/nginx/nginx.log info;
+
+    location / {
+        include uwsgi_params;
+        uwsgi_pass richo.tw:5000;
+    }
+}
