--- fig.yml
@@ -1,4 +1,0 @@
-web:
-  build: .
-  ports:
-   - "80:80"
