+++ requirements.txt
@@ -1,0 +1,2 @@
+flask==1.0.2
+uwsgi==2.0.17.1
