--- static/map/css/index.css
@@ -1,14 +1,0 @@
-body{
-	margin:0; padding:0;
-}
-
-#map {
-	position:absolute; top:0; bottom:0; width:100%;
-}
-
-.lineConnect {
-	fill: none;
-	stroke: #93FF93;
-	stroke-width: 2;
-	opacity: 1;
-}
