--- static/map/data/test.geojson
@@ -1,1 +1,0 @@
-{"type":"FeatureCollection","features":[{"type":"Feature","properties":{"marker-color":"#7e7e7e","marker-size":"medium","marker-symbol":"","id":"route"},"geometry":{"type":"Point","coordinates":[120.99715232849121,24.79666939032255]}},{"type":"Feature","properties":{"marker-color":"#7e7e7e","marker-size":"medium","marker-symbol":"","id":"route"},"geometry":{"type":"Point","coordinates":[120.9979462623596,24.796065530098083]}},{"type":"Feature","properties":{"marker-color":"#7e7e7e","marker-size":"medium","marker-symbol":"","id":"route"},"geometry":{"type":"Point","coordinates":[120.99839687347412,24.796494076366734]}},{"type":"Feature","properties":{"marker-color":"#7e7e7e","marker-size":"medium","marker-symbol":"","id":"route"},"geometry":{"type":"Point","coordinates":[121.0010576248169,24.79698105896543]}},{"type":"Feature","properties":{"marker-color":"#7e7e7e","marker-size":"medium","marker-symbol":"","id":"route"},"geometry":{"type":"Point","coordinates":[121.00163698196411,24.796922621154543]}},{"type":"Feature","properties":{"marker-color":"#7e7e7e","marker-size":"medium","marker-symbol":"","id":"route"},"geometry":{"type":"Point","coordinates":[121.0020875930786,24.796533035044995]}},{"type":"Feature","properties":{"marker-color":"#7e7e7e","marker-size":"medium","marker-symbol":"","id":"route"},"geometry":{"type":"Point","coordinates":[121.0029673576355,24.798052413951396]}},{"type":"Feature","properties":{"marker-color":"#7e7e7e","marker-size":"medium","marker-symbol":"","id":"route"},"geometry":{"type":"Point","coordinates":[121.0035252571106,24.80003926597158]}},{"type":"Feature","properties":{"marker-color":"#7e7e7e","marker-size":"medium","marker-symbol":"","id":"route"},"geometry":{"type":"Point","coordinates":[121.00818157196044,24.797760227145858]}},{"type":"Feature","properties":{"marker-color":"#7e7e7e","marker-size":"medium","marker-symbol":"","id":"route"},"geometry":{"type":"Point","coordinates":[121.01326704025269,24.795656461823903]}},{"type":"Feature","properties":{"marker-color":"#7e7e7e","marker-size":"medium","marker-symbol":"","id":"route"},"geometry":{"type":"Point","coordinates":[121.01418972015381,24.802045564628408]}},{"type":"Feature","properties":{"marker-color":"#7e7e7e","marker-size":"medium","marker-symbol":"","id":"route"},"geometry":{"type":"Point","coordinates":[121.02092742919922,24.809797753364574]}},{"type":"Feature","properties":{"marker-color":"#7e7e7e","marker-size":"medium","marker-symbol":"","id":"route"},"geometry":{"type":"Point","coordinates":[121.0232663154602,24.813089365258648]}},{"type":"Feature","properties":{"marker-color":"#7e7e7e","marker-size":"medium","marker-symbol":"","id":"route"},"geometry":{"type":"Point","coordinates":[121.02395296096802,24.81281669067142]}},{"type":"Feature","properties":{"marker-color":"#7e7e7e","marker-size":"medium","marker-symbol":"","id":"route"},"geometry":{"type":"Point","coordinates":[121.02378129959106,24.812544015484203]}}]}
