--- templates/map.html
@@ -1,25 +1,0 @@
-<!doctype html>
-<html>
-<head>
-	<meta charset="utf-8">
-	<meta http-equiv="X-UA-Compatible" content="IE=edge">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-
-	<script src='https://api.tiles.mapbox.com/mapbox.js/v2.1.5/mapbox.js'></script>
-	<link href='https://api.tiles.mapbox.com/mapbox.js/v2.1.5/mapbox.css' rel='stylesheet' />
-	<link href="{{ url_for('static', filename='map/css/index.css') }}" rel='stylesheet' />
-
-	<title>map4fun</title>
-
-</head>
-<body>
-
-	<div id='map'></div>
-
-	<script src='https://api.tiles.mapbox.com/mapbox.js/v2.1.5/mapbox.js'></script>
-	<script src="http://d3js.org/d3.v3.min.js" charset="utf-8"></script>
-	<script src="//code.jquery.com/jquery-1.11.2.min.js"></script>
-	<script src="{{ url_for('static', filename='map/js/index.js') }}"></script>
-
-</body>
-</html>
