--- webapp.ini
@@ -1,14 +1,0 @@
-[uwsgi]
-module = wsgi
-callable=app
-
-master = true
-processes = 5
-
-socket = webapp.sock
-chmod-socket = 660
-vacuum = true
-
-die-on-term = true
-
-logto = /home/richo/log/richo_site.log
