--- wsgi.py
@@ -1,4 +1,0 @@
-from webapp import app
-
-if __name__ == "__main__":
-    app.run()
